requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
sentencepiece==0.2.0
python-multipart==0.0.12
httpx==0.27.2
